requirements.txt
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_lambda
aws_cdk.aws_appsync
aws_cdk.aws_logs
aws_cdk.aws_dynamodb
aws_cdk.aws_iam
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws_cdk.aws_cloudfront
aws_cdk.aws_certificatemanager
